Prompt 10
Store the chosen option in a variable in the localStorage.
If the user changes their option, change the local storage variable.
Don't allow the user to vote more than once during 5 minutes.
Store the timestamp of the last vote in the localStorage.
When clicking the button, check if 5 minutes have passed by comparing with the timestamp in the localStorage.
After 5 minutes, allow the user to vote again or change their vote.