Prompt 11
I want to make this voting system p2p so the votes computing happens only in the clients, without needing a server. But before doing that, let's add a new admin.html page that will be the first peer with a constant ID used by the others to connect. Add a dashboard to the admin page where we can track the votes. Also, add an area for the log of peer events. Again, don't add the p2p functionality now; we will do it later.